requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
Pillow>=10.0.0

# File Monitoring
watchdog>=3.0.0

# Vector Database
chromadb>=0.4.0

# LLM Integration
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Machine Learning & Analytics
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment Management
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Date & Time
python-dateutil>=2.8.0

# PDF Generation for Reports
reportlab>=4.0.0

# Additional ML Libraries for Anomaly Detection
pyod>=1.1.0

# Image Processing for Fraud Detection
opencv-python>=4.8.0

# PDF to Image Conversion
pdf2image>=1.16.0

# Plotting and Visualization
plotly>=5.17.0

# Excel support
openpyxl>=3.1.0
